Refactor: Remove obsolete classes and files
- Deleted LLMASM.java and MMInstance.java as they are no longer needed.
- Removed NativeAccess.java and replaced it with a new implementation in intergrations/NativeAccess.java using JNR-FFI for better performance and compatibility.
- Cleared out test.masm and tui.lua files as they are outdated and not in use.
