--- Header.txt
@@ -1,27 +1,0 @@
-Header:
-- Magic Number: 0x4D42434D
-- Version: 1
-- Number of Instructions: 3
-
-Instructions:
-1. Opcode: 0x01 (MOV)
-   Operand 1 Type: 0x02 (Register)
-   Operand 1 Value: 0x00 (RAX)
-   Operand 2 Type: 0x06 (Immediate)
-   Operand 2 Value: 10
-2. Opcode: 0x02 (ADD)
-   Operand 1 Type: 0x02 (Register)
-   Operand 1 Value: 0x00 (RAX)
-   Operand 2 Type: 0x02 (Register)
-   Operand 2 Value: 0x01 (RBX)
-3. Opcode: 0xFF (HLT)
-   Operand 1 Type: 0x00 (None)
-   Operand 1 Value: 0
-   Operand 2 Type: 0x00 (None)
-   Operand 2 Value: 0
-
-Label Map:
-- Number of Labels: 1
-- Label 1:
-  - Name: "main"
-  - Address: 0
