--- dead.masm
@@ -1,12 +1,0 @@
-lbl main
-inc RAX
-inc RAX
-inc RAX
-inc RAX
-out 1 RAX
-
-jmp #end
-inc RAX
-
-lbl end
-hlt
