--- my_module.ll
@@ -1,7 +1,0 @@
-; ModuleID = 'my_module'
-source_filename = "my_module"
-
-define i32 @my_function() {
-entry:
-  ret i32 0
-}
