--- src/main/java/org/finite/ModuleManager/MMInstance.java
@@ -1,14 +1,0 @@
-package org.finite.ModuleManager;
-
-
-// ModuleManager For finite's MASM Interpreter
-
-/*
-    * ModuleManager is a class that manages the modules that are loaded into the interpreter
-    * these are things like plugins or even full desktop environments
-    * the ModuleManager is responsible for loading and unloading these modules aswell as handling the dependencies and the order in which they are loaded
-    * the ModuleManager is also responsible for handling the module's lifecycle, such as starting and stopping the module
-    * the ModuleManager is also responsible for handling the module's configuration, such as the configuration file, and the configuration directory
- */
-
-
