--- test.masm
@@ -1,6 +1,0 @@
-lbl main
-out 1 RAX
-cout 1 10
-inc RAX
-cmp RAX 100000
-jne #main
